refactor: util classes moved into global package and renamed
